Remove ByteBuddy dependency and related configuration from pom.xml
The ByteBuddy dependency was used by the surefire plugin to allow inline-mockito
to be run as java agent.

Unfortunately, this definition broke the build in case of a fresh local maven
repository, because some project modules implement tests without Mockito usage,
in which case the ByteBuddy dependency is not downloaded and present at the time
the tests execute. Surefire tries to start the JVM nevertheless with ByteBuddy
as java agent, which fails.

Note, that this problem occurs only, if any module with unit tests and no
Mockito dependency is compiled and tested before any module with Mockito
dependency, which makes building even more brittle.

This commit removes the ByteBuddy dependency to fix the problem.